src/requirements.txt
Flask==2.1.0
python-dotenv==0.21.0
beaker-pyteal==0.4.1